requirements.txt
requests==2.22.0
daemonize==2.5.0
psutil==5.7.0
Django==3.0.8
djangorestframework==3.11.0